1. rename extractor 2. add idea ignore
